temp

+++ algorithm_C++/layout/base_optimizer.hpp
@@ -12,8 +12,8 @@ public:
     virtual ~BaseOptimizer() {}
     virtual mat::Mat optimize(mat::Mat &initial_x) {}
     virtual void initialize(mat::Mat, int, mat::Mat *w=NULL) {}
-    virtual float compute_stress(mat::Mat x) {}
-//    virtual float compute_stress(const mat::Mat&) {}
+    virtual double compute_stress(mat::Mat x) {}
+//    virtual double compute_stress(const mat::Mat&) {}
 };
 
 #endif //MULTILEVEL_STRESS_C___BASECLASS_H
