+++ algorithm_C++/layout/sgd/sgd.cpp
@@ -23,9 +23,9 @@ void SGDOptimizer::initialize(mat::Mat target_dist, int dim, mat::Mat *w) {
     target_dim = dim;
 }
 
-float SGDOptimizer::compute_stress(mat::Mat x){
-    float stress = 0;
-    float stress_ij;
+double SGDOptimizer::compute_stress(mat::Mat x){
+    double stress = 0;
+    double stress_ij;
     for (int i = 0; i < n_nodes; ++i){
         for (int j = i + 1; j < n_nodes; ++j){
             stress_ij = (x[i] - x[j]).l2_norm() / dist(i, j) - 1;
@@ -37,21 +37,21 @@ float SGDOptimizer::compute_stress(mat::Mat x){
     return stress;
 }
 
-void SGDOptimizer::get_etas(std::vector<float> &etas, int t_max, int t_maxmax, float eps) {
-    float min_weight = POS_INF;
-    float max_weight = NEG_INF;
+void SGDOptimizer::get_etas(std::vector<double> &etas, int t_max, int t_maxmax, double eps) {
+    double min_weight = POS_INF;
+    double max_weight = NEG_INF;
     for (int i = 0; i < n_nodes; ++i){
         for (int j = i + 1; j < n_nodes; ++j){
             min_weight = std::min(min_weight, weights(i, j));
             max_weight = std::max(max_weight, weights(i, j));
         }
     }
-    float eta_max = 1 / min_weight;
-    float eta_min = eps / max_weight;
-    float l = std::log(eta_max / eta_min) / (t_max - 1);
+    double eta_max = 1 / min_weight;
+    double eta_min = eps / max_weight;
+    double l = std::log(eta_max / eta_min) / (t_max - 1);
 
-    float eta_switch = 1 / max_weight;
-    float eta;
+    double eta_switch = 1 / max_weight;
+    double eta;
     for (int t = 0; t < t_maxmax; ++t){
         eta = eta_max * std::exp(-l * t);
         if (eta < eta_switch) break;
@@ -65,14 +65,14 @@ void SGDOptimizer::get_etas(std::vector<float> &etas, int t_max, int t_maxmax, f
     }
 }
 
-float SGDOptimizer::optimize_iter(mat::Mat &pos, std::vector<std::pair<int, int>> &all_pairs, float eta) {
+double SGDOptimizer::optimize_iter(mat::Mat &pos, std::vector<std::pair<int, int>> &all_pairs, double eta) {
     /*
      * 一轮更新过程，会直接修改传入的pos对象
      */
     std::random_shuffle(all_pairs.begin(), all_pairs.end());
     int src, dst;
-    float w_ij, d_ij, mu, dx, dy, mag, delta, r, r_x, r_y;
-    float max_delta = 0;
+    double w_ij, d_ij, mu, dx, dy, mag, delta, r, r_x, r_y;
+    double max_delta = 0;
     for (std::pair<int, int> &p : all_pairs){
         src = p.first;
         dst = p.second;
@@ -81,7 +81,7 @@ float SGDOptimizer::optimize_iter(mat::Mat &pos, std::vector<std::pair<int, int>
         d_ij = dist(src, dst);
 
         mu = eta * w_ij;
-        mu = std::min(mu, float(1));
+        mu = std::min(mu, double(1));
 
         dx = pos(src, 0) - pos(dst, 0);
         dy = pos(src, 1) - pos(dst, 1);
@@ -103,7 +103,7 @@ float SGDOptimizer::optimize_iter(mat::Mat &pos, std::vector<std::pair<int, int>
 }
 
 mat::Mat SGDOptimizer::optimize(mat::Mat &initial_x) {
-    float stop_th = 0.03;  // 当所有节点的最大移动距离小于这个值时算法停止
+    double stop_th = 0.03;  // 当所有节点的最大移动距离小于这个值时算法停止
     // 生成所有节点对
     std::vector<std::pair<int, int>> all_pairs;
     for (int src = 0; src < n_nodes; ++src){
@@ -113,11 +113,11 @@ mat::Mat SGDOptimizer::optimize(mat::Mat &initial_x) {
     }
 
     // 生成所有的eta
-    std::vector<float> etas;
+    std::vector<double> etas;
     get_etas(etas);
 
-    float max_delta;
-    for (float eta : etas){
+    double max_delta;
+    for (double eta : etas){
         max_delta = optimize_iter(initial_x, all_pairs, eta);
         std::cout << "stress: " << compute_stress(initial_x) << " max delta: " << max_delta << std::endl;
         if (max_delta < stop_th) break;
