+++ algorithm_C++/multiscale/adapted_init.h
@@ -16,12 +16,12 @@
 #include<ctime>
 
 namespace adapted_init{
-    const float lr_max = 10;
-    const float lr_min = 10;
-    const float max_step = 40;
+    const double lr_max = 10;
+    const double lr_min = 10;
+    const double max_step = 40;
     const int max_effective_center_dist = 20;  // stress迭代更新时考虑的最远的center距离
     const int th = 100;
-    const float ratio = 0.33;
+    const double ratio = 0.33;
 
     struct dist_pair{
         int target_center;
@@ -37,7 +37,7 @@ namespace adapted_init{
     };
 
     void kcenter(std::vector<int> &centers, std::vector<std::set<dist_pair>> &dist_of_centers, int k, mat::Mat &shortest_dist);
-    float get_lr(int step_cnt);
+    double get_lr(int step_cnt);
     mat::Mat solve_r(BaseOptimizer *optimizer, mat::Mat dist, mat::Mat weights, int target_dim, std::vector<int> &k_list, int k_idx);
     void generate_k_list(std::vector<int> &k_list, int n_nodes, int ratio, int th) ;
     mat::Mat solve(BaseOptimizer *optimizer, Graph &graph, int target_dim);
