+++ algorithm_C++/test_cuda/a.cu
@@ -7,7 +7,7 @@
 #define BLOCK_Y 4
 #define N 8
 
-__global__ void add_(float *a, float *b, float *c){
+__global__ void add_(double *a, double *b, double *c){
     int i = threadIdx.x + BLOCK_X * blockIdx.x;
     int j = threadIdx.y + BLOCK_Y * blockIdx.y;
     int idx = i * N + j;
@@ -16,13 +16,13 @@ __global__ void add_(float *a, float *b, float *c){
 
 ClassA::ClassA(int size) {
     n = size;
-    cudaMalloc((void**)&d_data, sizeof(float) * size);
+    cudaMalloc((void**)&d_data, sizeof(double) * size);
 }
 
 ClassA::ClassA(ClassA &&other) {
     n = other.n;
-    cudaMalloc((void**)&d_data, sizeof(float) * n);
-    cudaMemcpy(d_data, other.d_data, sizeof(float) * n, cudaMemcpyDeviceToDevice);
+    cudaMalloc((void**)&d_data, sizeof(double) * n);
+    cudaMemcpy(d_data, other.d_data, sizeof(double) * n, cudaMemcpyDeviceToDevice);
 }
 
 ClassA ClassA::add(ClassA &other) {
@@ -40,17 +40,17 @@ ClassA::~ClassA() {
 }
 
 void ClassA::to_host() {
-    data = (float*)malloc(sizeof(float) * n);
-    cudaMemcpy(data, d_data, sizeof(float) * n, cudaMemcpyDeviceToHost);
+    data = (double*)malloc(sizeof(double) * n);
+    cudaMemcpy(data, d_data, sizeof(double) * n, cudaMemcpyDeviceToHost);
 }
 
 ClassA random(int n){
     ClassA ans(n);
-    float *x = (float*)malloc(sizeof(float) * n);
+    double *x = (double*)malloc(sizeof(double) * n);
     for (int i = 0; i < n; ++i){
         x[i] = rand() % 10;
         cout << x[i] << endl;
     }
-    cudaMemcpy(ans.d_data, x, sizeof(float) * n, cudaMemcpyHostToDevice);
+    cudaMemcpy(ans.d_data, x, sizeof(double) * n, cudaMemcpyHostToDevice);
     return ans;
 }
